# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ti Saúde API",
        "description": "Documentação da API do desafio fullstack",
        "version": "1.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/consultas": {
            "get": {
                "tags": [
                    "Consultas"
                ],
                "summary": "Lista todas as consultas",
                "operationId": "6b9a9deca0de3bd3c0ce4e567c2bc307",
                "responses": {
                    "200": {
                        "description": "Consultas listadas com sucesso"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Consultas"
                ],
                "summary": "Cria uma nova consulta",
                "operationId": "8a267506eaf2ad9fbd69dd08002284a9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "data",
                                    "hora",
                                    "paciente_id",
                                    "medico_id"
                                ],
                                "properties": {
                                    "cons_codigo": {
                                        "type": "string",
                                        "example": "CONS001"
                                    },
                                    "data": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-04-22"
                                    },
                                    "hora": {
                                        "type": "string",
                                        "example": "14:30"
                                    },
                                    "particular": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "paciente_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "medico_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "plano_saude_id": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "procedimentos": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Consulta criada com sucesso"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/consultas/{id}": {
            "get": {
                "tags": [
                    "Consultas"
                ],
                "summary": "Busca uma consulta por ID",
                "operationId": "a822674ac6c730b13aba834afb2ca4f3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Consulta encontrada com sucesso"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "404": {
                        "description": "Consulta não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Consultas"
                ],
                "summary": "Atualiza uma consulta",
                "operationId": "31bcc74db8e153d2a0cd9e2d2b106778",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "data",
                                    "hora",
                                    "paciente_id",
                                    "medico_id"
                                ],
                                "properties": {
                                    "cons_codigo": {
                                        "type": "string",
                                        "example": "CONS001"
                                    },
                                    "data": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-04-22"
                                    },
                                    "hora": {
                                        "type": "string",
                                        "example": "14:30"
                                    },
                                    "particular": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "paciente_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "medico_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "plano_saude_id": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "procedimentos": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Consulta atualizada com sucesso"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "404": {
                        "description": "Consulta não encontrada"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Consultas"
                ],
                "summary": "Remove uma consulta",
                "operationId": "d95d7edacdc5ea30c86750bd6ef9ae95",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Consulta removida com sucesso"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "404": {
                        "description": "Consulta não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}